+++ packages/gestalt-codemods/1.22.0-1.23.0/__testfixtures__/toast-remove-color-icon-single.input.js
@@ -1,0 +1,9 @@
+// @flow
+import React from 'react';
+import { Toast } from 'gestalt';
+
+export default function ToastExample() {
+  return (
+    <Toast text="Help" color="orange" />
+  );
+}